fix(lfs): compile error with 32 bit build
When compiling for a 32 bit architecture the size of type 'size_t'
changes and there are different pre compiler macros for printf() format
parameter used on Unix and Microsoft Windows platforms. With rework of
32 bit builds I find theese limitations that must be fixed:

* Limitation: compile 32 bit on 64 bit MS Windows isn't possible because
              pthreads4w only compile to host architecture. Pthreads4w
              needs NMake Makefiles and that does not support platform
              specification. To get 32 bit libraries you must compile on
              a host with 32 bit MS Windows platform.

* Limitation: MS Windows build with OpenSSL segfaults due to outdated
              OpenSSL prebuilt libraries.

* Limitation: Linux 32 bit build with OpenSSL fails due to 64 bit
              OpenSSL prebuilt libraries.

Also rename the sample subdirectory 'sample/' to first letter upper case
'Sample/' to follow convention of other "main" subdirectories.