re-en(addressing): doing some progam maintenance
Add getter 'CNetadapter::prefix_length()' to get the length of the
network prefix from an IP address instead of only the netmask.

Improve logging messages when starting and register an UPnP device.

Doing some documentation.

Bugfix in 'SSockaddr::netaddr()'. A buffer was too small to hold a
netaddress with scope id.