fix(addressing): fix binding for listening
Three hard coded atempts are made to bind() a socket to a local network
address if it fails.

Random port number generator is removed. Also the algorithm to repeat
failed binding with different socket file descriptors. That is not the
issue.

Don't set scope id when binding a GUA for ssdp. This isn't accepted by
'::bind()' on win32.