fix(addressing): bug CNetadapter with find address
I assumed that the access to the internal network adapter list is
sequencial in adapter and then in its ip addresses. That is not true. A
node will choose the source address for a communication following some
rules of best choice, per RFC3484. For example I have seen that the
operating system returned first a link local address on adapter 2, then
a global unicast address on adapter 3, and again from adapter 2 an IPv4
address. You cannot expect any order on adapter and/or ip address.

Now I use a simple state machine to match all steps from find_first() to
find_next().