fix(addressing): bugs in CNetadapter class
Disabling umock::* injection in CNetadapter class due to segfaults in
Unit Tests on destruction of this class. It's because umock injection is
not under control of the class and that violates the RAII principle.

Mock Unit Tests for CNetadapter and create additional tests to mock
different local network adapter ip address combinations. Fix CNetadapter
class with help of the mocking tests.

Reworking of compatible free function 'UpnpGetIfInfo()' to fit it to the
modified CNetadapter class and moving it to the compa:: namespace.
Modifying and enable some Unit Tests for 'UpnpGetIfInfo()'.

Modifying documentation for the CNetadapter class to fit to the new
behavior.

By the way:
Changing argument 'a_prefixlength' of free function
'bitmask_to_netmask()' from type 'const uint8_t' to 'const unsigned int'
to be portable.