re-en(addressing): rework tests for UpnpGetIfInfo
Useing real local network adapter infos for testing.
Disabling umock::* injection in CNetadapter class due to segfaults in
Unit Tests on destruction of this class. It's because umock injection is
not under control of the class and that violates the RAII principle.

Fixing bug in CNetadapter::find_next(). It has only found one loopback
ip address on a local network adapter with more than one ip address,
e.g. "[::1]" and "127.0.0.1".