fix(addressing): solve problem w/ ABSOLUTE symbol
On MS Windows there are already constants ABSOLUTE and RELATIVE defined.
These are also defined in the pupnp sdk but with other meaning. This is
no problem on Unix like platforms. But on Microsoft Windows this should
be complained by the compiler due to type missmatch but in the program
this error is hidden by type cast.