build(project): Improve test API calls
Testing the Application Programming Interface is performed with bash
build- and test-scripts to be independent from Googletest and to avoid
edge conditions with CMake and github. Optimizing the API is mainly done
with conditional compiling of the different program modules. For testing
with original "old pupnp" libraries (libupnp, ixml) its binary libraries
are provided in Apps/bin.

More tests are enabled now.