re-en(addressing): Remove url-classes and -parsers
Remove Curi class. Using it is of no use. Its methods are only redirects
to system calls.
Remove unused and unfinished uri/url code, that is based on the "URL
Living Standard". Most of it can be done with system call
'getaddrinfo()'. But it has some useful details that may be of interest
later, that are:
* C++ character parsing
* Redirect std::clog stream to be used for thread safe program messages
* Using a Finite State Machine (fsm)